Move sync from model to task
closes #3260
https://pulp.plan.io/issues/3260
